postList moved to posts folder, userList moved to users folder
